fixing labels second screen
