+++ learningCodableJSON/ViewController.swift
@@ -17,7 +17,8 @@ class ViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-            print(getJSON())
+            
+            
     }
 }
